Connected the backend to the chart to get the data from the database

+++ src/main/java/com/SDP/project/Repository/InventoryRepository.java
@@ -2,11 +2,13 @@ package com.SDP.project.Repository;
 
 import com.SDP.project.models.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface InventoryRepository extends JpaRepository<Inventory, Long> {
+public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
     Optional<Inventory> findByTaskId(int taskId);
 }
